notification_service/requirements.txt
Django>=4.0
gunicorn
psycopg2-binary
dj-database-url
djangorestframework
djangorestframework-simplejwt
python-decouple
cryptography
requests
django-cors-headers
pika
